--- apps/nuxt-app/utils/auth/index.ts
@@ -1,6 +1,0 @@
-export * from './authHelpers';
-export * from './authClient';
-export * from './authClientDirect';
-export * from './authClientProxy';
-export * from './authService';
-export * from './authServiceClient';
