+++ packages/sdk/crm/src/auth/authClient.ts
@@ -10,6 +10,53 @@ export abstract class AuthClient {
     this.cookieService = new CookieService();
   }
 
+  abstract login(
+    credentials: AuthCredentials,
+  ): Promise<AuthResponse | undefined>;
+
+  abstract refresh(): Promise<AuthResponse | undefined>;
+  abstract changePassword(
+    credentials: AuthCredentials,
+  ): Promise<AuthResponse | undefined>;
+
+  abstract logout(): Promise<boolean>;
+
+  abstract getUser(): Promise<AuthResponse | undefined>;
+
+  protected async getUserFromCookie(
+    token?: string,
+    refreshToken?: string,
+  ): Promise<AuthResponse | undefined> {
+    token = token || this.getCookieUserToken();
+    refreshToken = refreshToken || this.getCookieRefreshToken();
+    if (!token || !refreshToken) {
+      return undefined;
+    }
+
+    return AuthService.getUserObjectFromToken(token, refreshToken);
+  }
+
+  // get cookie values
+  protected getCookieUser(): string {
+    return this.cookieService.get(AUTH_COOKIES.USER);
+  }
+
+  protected getCookieRefreshToken(): string {
+    return this.cookieService.get(AUTH_COOKIES.REFRESH_TOKEN);
+  }
+
+  protected getCookieUserToken(): string {
+    return this.cookieService.get(AUTH_COOKIES.USER_AUTH);
+  }
+
+  protected getCookieMaxAge(): number {
+    const maxAge = this.cookieService.get(AUTH_COOKIES.USER_MAX_AGE);
+    if (!maxAge) {
+      return 1800;
+    }
+    return parseInt(maxAge, 10);
+  }
+
   public spoofPreviewUser(token: string): void {
     this.clearCookies();
     const maxAge = 1800;
@@ -64,60 +111,7 @@ export abstract class AuthClient {
     });
   }
 
-  abstract login(credentials: AuthCredentials): Promise<AuthResponse>;
-
-  abstract logout(): Promise<boolean>;
-
-  abstract refresh(): Promise<AuthResponse | undefined>;
-
-  async getUser(
-    token?: string,
-    refreshToken?: string,
-  ): Promise<AuthResponse | undefined> {
-    const result = await this.getUserFromCookie(token, refreshToken);
-    if (!result) {
-      return undefined;
-    }
-
-    if (result && result.succeeded && result.tokens?.token) {
-      if (result.tokens.expired) {
-        this.clearCookies();
-        return undefined;
-      } else if (result.tokens.expiresSoon) {
-        await this.refresh();
-      }
-    } else {
-      this.clearCookies();
-    }
-    return result;
-  }
-
-  protected async getUserFromCookie(
-    token?: string,
-    refreshToken?: string,
-  ): Promise<AuthResponse | undefined> {
-    token = token || this.getCookieUserToken();
-    refreshToken = refreshToken || this.getCookieRefreshToken();
-    if (!token || !refreshToken) {
-      return undefined;
-    }
-
-    return AuthService.getUserObjectFromToken(token, refreshToken);
-  }
-
-  public async changePassword(
-    credentials: AuthCredentials,
-  ): Promise<AuthResponse> {
-    if (!credentials.newPassword) {
-      throw new Error('New password is required');
-    }
-    return await this.changePasswordImplementation(credentials);
-  }
-
-  protected abstract changePasswordImplementation(
-    credentials: AuthCredentials,
-  ): Promise<AuthResponse>;
-
+  // set cookie values
   protected setCookiesLogin(
     authResponse: AuthResponse,
     rememberUser: boolean,
@@ -167,6 +161,7 @@ export abstract class AuthClient {
       // httpOnly: true,
     });
   }
+
   protected setCookieUserToken(token: string, maxAge?: number): void {
     if (!maxAge) {
       maxAge = this.getCookieMaxAge();
@@ -176,22 +171,6 @@ export abstract class AuthClient {
       payload: token,
       maxAge,
     });
-  }
-
-  protected getCookieRefreshToken(): string {
-    return this.cookieService.get(AUTH_COOKIES.REFRESH_TOKEN);
-  }
-
-  protected getCookieUserToken(): string {
-    return this.cookieService.get(AUTH_COOKIES.USER_AUTH);
-  }
-
-  protected getCookieMaxAge(): number {
-    const maxAge = this.cookieService.get(AUTH_COOKIES.USER_MAX_AGE);
-    if (!maxAge) {
-      return 1800;
-    }
-    return parseInt(maxAge, 10);
   }
 
   public clearCookies(): void {
